Make Download MultiTypeComponent
